Adding the ability of starting to fingerprint at a variable position in time represented by doubles instead of ints.

+++ src/SoundFingerprinting.Audio.NAudio/NAudioService.cs
@@ -29,7 +29,7 @@
             }
         }
 
-        public override AudioSamples ReadMonoSamplesFromFile(string pathToSourceFile, int sampleRate, int seconds, int startAt)
+        public override AudioSamples ReadMonoSamplesFromFile(string pathToSourceFile, int sampleRate, double seconds, double startAt)
         {
             float[] samples = sourceReader.ReadMonoFromSource(pathToSourceFile, sampleRate, seconds, startAt);
             return new AudioSamples
